06-google-cloud-function-with-real-case-studies/code/Demo2 - Cloud  Firestore & Pubsub/requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-pubsub==2.8.0